Rename: {Sequential->Local}SimulationTask, move it around
